requirements.txt
contourpy
gensim
keras
keras-preprocessing
matplotlib
nltk
opencv-contrib-python
pandas
pillow
pydot
scikit-learn
scipy
seaborn
tensorflow==2.10.1
torchaudio
torchdata
pillow
torchvision==0.14.1
torchviz==0.0.2

